review(mesh): F4 -- restructure F3 pin tests + close 2 CodeQL findings
Two parallel changes addressing reviewer + CodeQL feedback:

## 1. CodeQL findings closed (alerts #256, #257)

* `audit.py:844` -- unused `sig = None` in the widened fail-soft
  except branch. The variable is initialised to `None` at the top
  of the block; the assignment in the `except Exception` branch was
  redundant (control falls through to the write block which only
  attaches `sig` when non-None via the `else` branch). Removed
  the redundant assignment with a clarifying comment.

* `tests/mesh/test_pr195_review_followups.py:77` -- unused
  `StubRobot` class in TestPermissiveACLWarningExceptNarrow.
  The test is a static source-code assertion; the Mesh-construction
  scaffolding never got exercised. Removed the dead class.

## 2. Test reorganisation: F3 pin tests moved to per-subject homes

The omnibus `test_pr195_review_followups.py` was bad naming -- per
the in-tree convention, test files should describe **what** is
tested, not **when** it was added. R16 (`c3627e8`) fixed exactly
the same problem for two other files; this PR reintroduced it.

Distributed the F3 pin tests into their proper homes:

| Class                                  | New home                                    |
|----------------------------------------|---------------------------------------------|
| TestJSON5DepSwap                       | test_acl_config.py (alongside other JSON5)  |
| TestPermissiveACLWarningExceptNarrow   | test_default_acl_warning.py                 |
| TestCameraDisabledLenientParse         | test_camera_privacy_kill_switch.py (NEW)    |
| TestResumeStrictPeerId                 | test_resume_replay.py                       |
| TestEstopRedundantAudit                | test_estop_replay.py                        |
| TestPSKStateLock + TestAuditFailSoft   | test_audit_integrity.py                     |
| TestVerifyCaPinSymlink                 | test_iot_ca_pin.py                          |
| TestPolicyHostAllowlistValidation      | test_policy_host_allowlist.py (NEW)         |

Two new files were created where no good per-subject home existed:
`test_camera_privacy_kill_switch.py` (STRANDS_MESH_CAMERA_DISABLED
parsing) and `test_policy_host_allowlist.py` (STRANDS_MESH_POLICY_
HOST_ALLOW operator-entry validation). Both follow the existing
naming convention.

The omnibus `test_pr195_review_followups.py` is deleted; the same
24 pin tests now live alongside the rest of their per-subject
suite, which is the convention the test_zenoh_transport_security
rename established (R16).

## Verification

* 798 passed, 4 skipped (no test count change; same coverage)
* lint clean, format clean
* CodeQL findings #256 + #257 closed (the file they referenced is
  gone for #257; the unused-variable for #256 is removed)
